replace LocalBroadcastManager with direct notification sending
