+++ android/app/src/main/kotlin/io/rebble/fossil/WatchService.kt
@@ -4,23 +4,16 @@ import android.Manifest
 import android.app.Service
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothDevice
-import android.content.BroadcastReceiver
-import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
 import android.content.pm.PackageManager
 import android.os.Binder
 import android.os.Build
 import android.os.IBinder
-import android.provider.Telephony
 import android.widget.Toast
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
-import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import io.flutter.Log
 import io.rebble.libpebblecommon.ProtocolHandler
-import io.rebble.libpebblecommon.blobdb.NotificationSource
-import io.rebble.libpebblecommon.blobdb.PushNotification
 import io.rebble.libpebblecommon.services.notification.NotificationService
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
@@ -51,28 +44,6 @@ class WatchService : Service() {
         return pBinder
     }
 
-    private fun packageToSource(pkg: String?): NotificationSource {
-        //TODO: Check for other email clients
-        return when (pkg) {
-            "com.google.android.gm" -> NotificationSource.Email
-            "com.facebook.katana" -> NotificationSource.Facebook
-            "com.twitter.android", "com.twitter.android.lite" -> NotificationSource.Twitter
-            Telephony.Sms.getDefaultSmsPackage(this) -> NotificationSource.SMS
-            else -> NotificationSource.Generic
-        }
-    }
-
-    @ExperimentalStdlibApi
-    private val notifBroadcastReceiver = object : BroadcastReceiver() {
-        override fun onReceive(context: Context?, intent: Intent?) {
-            val notif = intent?.getBundleExtra("notification") ?: return
-            //TODO centralize coroutine creation
-            GlobalScope.launch {
-                val res = notificationService.send(PushNotification(notif["subject"] as String, notif["sender"] as String, notif["content"] as String, packageToSource(notif["pkg"] as String)))
-                Log.d("FossilNotifRet", "Got resp from BlobDB: ${res::class.simpleName}")
-            }
-        }
-    }
 
     @ExperimentalStdlibApi
     override fun onCreate() {
@@ -96,7 +67,6 @@ class WatchService : Service() {
         if (!bluetoothAdapter.isEnabled) {
             Log.w(logTag, "Bluetooth - Not enabled")
         }
-        LocalBroadcastManager.getInstance(applicationContext).registerReceiver(notifBroadcastReceiver, IntentFilter("io.rebble.fossil.NOTIFICATION_BROADCAST"))
 
         startIO()
     }
